refactor token tasks

+++ src/js/api/auth/delete-logging-out.mjs
@@ -1,11 +1,13 @@
 import { deleteItem } from "../../storage/localStorage/delete.mjs";
+import { loadItem } from "../../storage/localStorage/load.mjs";
 
-export function deleteStorageSigningOut() {
-   document.addEventListener("DOMContentLoaded", () => {
+export function deleteLoggingOut() {
+   const tokenExist = loadItem("token");
+   if (tokenExist) {
       const logoutBtn = document.querySelector("#logout");
       logoutBtn.addEventListener("click", () => {
          deleteItem("token");
          deleteItem("profile");
       });
-   });
+   }
 }
